refactor: change tape test into jest test (hub)
