+++ hub/package.json
@@ -36,6 +36,7 @@
     "@types/express-winston": "^4.0.0",
     "@types/fetch-mock": "^7.3.5",
     "@types/fs-extra": "^9.0.13",
+    "@types/jest": "^28.1.3",
     "@types/nanoid": "^3.0.0",
     "@types/node": "^18.0.0",
     "@types/node-fetch": "^2.6.2",
@@ -43,18 +44,16 @@
     "@types/tape": "^4.13.2",
     "@typescript-eslint/eslint-plugin": "^5.29.0",
     "@typescript-eslint/parser": "^5.29.0",
-    "c8": "^7.11.3",
     "cross-env": "^7.0.3",
     "eslint": "^8.18.0",
     "eslint-import-resolver-typescript": "^2.7.1",
     "eslint-plugin-import": "^2.26.0",
-    "esmock": "^1.7.5",
     "fetch-mock": "^9.11.0",
+    "jest": "^28.1.1",
     "npm-run-all": "^4.1.5",
     "source-map-support": "^0.5.21",
     "supertest": "^6.2.3",
-    "tape": "^5.5.3",
-    "tape-promise": "^4.0.0",
+    "ts-jest": "^28.0.5",
     "ts-node": "^10.8.1",
     "typescript": "^4.7.4",
     "typescript-json-schema": "^0.53.1"
@@ -69,8 +68,7 @@
     "build:schema": "typescript-json-schema tsconfig.json HubConfig --required --noExtraProps --refs=false -o config-schema.json",
     "lint": "eslint --ext .ts ./src -f unix",
     "lint:fix": "eslint --ext .ts ./src --fix",
-    "test:build": "tsc -p ./test/tsconfig.json",
-    "test:coverage": "cross-env NODE_ENV=test LOCAL_DRIVER_CONFIG_TEST_DATA=configs/drivers.json c8 ts-node test/src/index.ts",
+    "test:coverage": "cross-env NODE_ENV=test LOCAL_DRIVER_CONFIG_TEST_DATA=configs/drivers.json jest --coverage",
     "test": "npm-run-all build --parallel lint test:coverage",
     "start": "node lib/index.js"
   },
@@ -83,5 +81,18 @@
     "Jack Zampolin (jack@blockstack.com)",
     "Jude Nelson (jude@blockstack.com)"
   ],
+  "jest": {
+    "preset": "ts-jest",
+    "globals": {
+      "ts-jest": {
+        "useESM": true
+      }
+    },
+    "moduleNameMapper": {
+      "^(\\.{1,2}/.*)\\.js$": "$1"
+    },
+    "testTimeout": 30000,
+    "verbose": true
+  },
   "type": "module"
 }
