+++ hub/test/src/testDrivers/global.ts
@@ -1,0 +1,4 @@
+type DataMap = {key: string, data: string, etag: string}[];
+export const azDataMap: DataMap = []
+export const s3DataMap: DataMap = []
+export const gcDataMap: DataMap = []
